requirements.txt
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.0
biopython==1.84
flask==3.0.3 